Move the Chain type to xp-assets-registrar
