+++ primitives/assets-registrar/Cargo.toml
@@ -5,10 +5,13 @@ authors = ["The ChainX Authors"]
 edition = "2018"
 
 [dependencies]
+codec = { package = "parity-scale-codec", version = "1.3.4", features = ["derive"], default-features = false }
+serde = { version = "1.0", features = ["derive"], optional = true }
 impl-trait-for-tuples = "0.1.3"
 
 # Substrate primitives
 sp-runtime = { version = "2.0.0", default-features = false }
+sp-std = { version = "2.0.0", default-features = false }
 
 # ChainX primitives
 chainx-primitives = { path = "../../primitives", default-features = false }
@@ -16,6 +19,8 @@ chainx-primitives = { path = "../../primitives", default-features = false }
 [features]
 default = ["std"]
 std = [
+    "codec/std",
+    "serde",
     # Substrate primitives
     "sp-runtime/std",
     # ChainX primitives
